+++ src/pmacct.h
@@ -238,8 +238,8 @@ struct pcap_callback_data {
   u_char * bmed_table; 
   u_char * biss_table; 
   struct pcap_device *device;
-  u_int16_t ifindex_in;
-  u_int16_t ifindex_out;
+  u_int32_t ifindex_in;
+  u_int32_t ifindex_out;
 };
 
 struct _protocols_struct {
